[Robot Model] further refactor the robot model framework
