+++ aerial_robot_control/include/aerial_robot_control/control/control_base.h
@@ -38,7 +38,7 @@
 
 #include <aerial_robot_control/flight_navigation.h>
 #include <aerial_robot_estimation/state_estimation.h>
-#include <aerial_robot_model/transformable_aerial_robot_model.h>
+#include <aerial_robot_model/model/aerial_robot_model.h>
 #include <ros/ros.h>
 #include <spinal/PwmInfo.h>
 #include <spinal/UavInfo.h>
