+++ aerial_robot_model/include/aerial_robot_model/model/aerial_robot_model_ros.h
@@ -2,7 +2,7 @@
 /*********************************************************************
  * Software License Agreement (BSD License)
  *
- *  Copyright (c) 2016, JSK Lab
+ *  Copyright (c) 2018, JSK Lab
  *  All rights reserved.
  *
  *  Redistribution and use in source and binary forms, with or without
@@ -33,19 +33,47 @@
  *  POSSIBILITY OF SUCH DAMAGE.
  *********************************************************************/
 
-#include <aerial_robot_model/aerial_robot_model.h>
+#pragma once
 
-using namespace aerial_robot_model;
+#include <aerial_robot_model/model/aerial_robot_model.h>
+#include <aerial_robot_model/AddExtraModule.h>
+#include <pluginlib/class_loader.h>
+#include <spinal/DesireCoord.h>
+#include <tf/tf.h>
+#include <tf2_ros/transform_broadcaster.h>
+#include <tf2_ros/static_transform_broadcaster.h>
 
-class MultirotorRobotModel : public RobotModel {
 
-public:
-  MultirotorRobotModel(bool init_with_rosparam = true, bool verbose = false, double fc_f_min_thre = 0, double fc_t_min_thre = 0, double epsilon = 10.0):
-    RobotModel(init_with_rosparam, verbose, fc_f_min_thre, fc_t_min_thre, epsilon)
-  {
-  }
-};
+namespace aerial_robot_model {
 
-/* plugin registration */
-#include <pluginlib/class_list_macros.h>
-PLUGINLIB_EXPORT_CLASS(MultirotorRobotModel, aerial_robot_model::RobotModel);
+  //Transformable Aerial Robot Model with ROS functions
+  class RobotModelRos {
+  public:
+    RobotModelRos(ros::NodeHandle nh, ros::NodeHandle nhp);
+    virtual ~RobotModelRos() = default;
+
+    //public functions
+    sensor_msgs::JointState getJointState() const { return joint_state_; }
+
+    const boost::shared_ptr<aerial_robot_model::RobotModel> getRobotModel() const { return robot_model_; }
+
+  private:
+    //private attributes
+    ros::ServiceServer add_extra_module_service_;
+    ros::Subscriber desire_coordinate_sub_;
+    ros::Subscriber joint_state_sub_;
+    tf2_ros::TransformBroadcaster br_;
+    tf2_ros::StaticTransformBroadcaster static_br_;
+    sensor_msgs::JointState joint_state_;
+    ros::NodeHandle nh_;
+    ros::NodeHandle nhp_;
+    pluginlib::ClassLoader<aerial_robot_model::RobotModel> robot_model_loader_;
+    boost::shared_ptr<aerial_robot_model::RobotModel> robot_model_;
+    std::string tf_prefix_;
+
+    //private functions
+    void jointStateCallback(const sensor_msgs::JointStateConstPtr& state);
+    bool addExtraModuleCallback(aerial_robot_model::AddExtraModule::Request& req, aerial_robot_model::AddExtraModule::Response& res);
+    void desireCoordinateCallback(const spinal::DesireCoordConstPtr& msg);
+  };
+} //namespace aerial_robot_model
