+++ aerial_robot_model/include/aerial_robot_model/model/transformable_aerial_robot_model.h
@@ -35,7 +35,7 @@
 
 #pragma once
 
-#include <aerial_robot_model/aerial_robot_model.h>
+#include <aerial_robot_model/model/aerial_robot_model.h>
 
 namespace aerial_robot_model {
   namespace transformable {
