+++ aerial_robot_model/include/aerial_robot_model/utils/math_utils.h
