+++ aerial_robot_simulation/include/aerial_robot_simulation/aerial_robot_hw_sim.h
@@ -40,7 +40,7 @@
 #ifndef _GAZEBO_ROS_CONTROL___AERIAL_ROBOT_HW_SIM_H_
 #define _GAZEBO_ROS_CONTROL___AERIAL_ROBOT_HW_SIM_H_
 
-#include <aerial_robot_model/transformable_aerial_robot_model.h>
+#include <aerial_robot_model/model/aerial_robot_model.h>
 #include <aerial_robot_simulation/noise_model.h>
 #include <aerial_robot_simulation/spinal_interface.h>
 #include <gazebo_ros_control/default_robot_hw_sim.h>
