+++ robots/hydrus/include/hydrus/hydrus_robot_model.h
@@ -35,7 +35,7 @@
 
 #pragma once
 
-#include <aerial_robot_model/transformable_aerial_robot_model.h>
+#include <aerial_robot_model/model/transformable_aerial_robot_model.h>
 
 class HydrusRobotModel : public aerial_robot_model::transformable::RobotModel {
 public:
